requirements.txt
Flask==0.12.2
Requests==2.18.4
PyGithub==1.43.5